requirements.txt
torch>=1.7.1
torchvision>=0.8.2
opencv-python>=4.5.0
numpy>=1.20.0
pillow>=8.0.0
ultralytics>=8.0.0
matplotlib>=3.3.0
pandas>=1.1.0
imutils>=0.5.4
tqdm>=4.60.0